fix(quotes): money columns are bigint, and the schema's invariants are tested
Two review findings on the quotes schema.

1) Every money column had been narrowed to `int`: quotes.total_minor,
   quote_items.unit_amount_minor/subtotal_minor and
   booking_custom_items.unit_amount_minor/subtotal_minor. 20260615121000 widened
   bookings.total_minor and booking_items' amounts to bigint under the header
   "int capped at ~21.4M EUR-cents; full-boat charters overflow" — and a bespoke
   quote is exactly that case, so the one module written for full-boat charters
   was the one place still capped. All five are bigint now, in the migration and
   in the byte-identical catch-up.sql block, with setup.sql regenerated to match.
   The comment justifying the narrowing claimed "bookings.total_minor is `int`",
   which is false at runtime; it is deleted rather than restated.

2) Nothing exercised the new schema. The ledger and catch-up parity guards only
   assert the file is mirrored — they would pass just the same with the tables
   spelled wrong. tests/integration/quotes-schema.test.ts (not tests/db/, which
   holds harness helpers only and has no test files) boots PGlite and asserts the
   invariants the migration exists to enforce, before api_convert_quote is built
   on top of them: bigint money columns, anon refused both reads and writes on
   all three tables, quotes.booking_id UNIQUE, quote_item_shape in both
   directions, and booking_custom_items refusing kind = 'catalogue'.

+++ supabase/migrations/20260909000000_quotes.sql
@@ -60,10 +60,11 @@ create table if not exists quotes (
   customer_phone text,
   status quote_status not null default 'draft',
   currency text not null default 'EUR',
-  -- Minor units, like every other money column here (bookings.total_minor is `int`, and this figure
-  -- is copied straight into it at conversion — same type, so a too-large quote fails when it is
-  -- entered, not silently at the insert).
-  total_minor int not null default 0,
+  -- Minor units, and `bigint` like every other money column on the money path: 20260615121000
+  -- widened bookings.total_minor and booking_items' amounts precisely because int caps at ~21.4M
+  -- EUR-cents and a full-boat charter overflows it. A bespoke quote IS that case, and this figure is
+  -- copied straight into bookings.total_minor at conversion, so it must be the same width.
+  total_minor bigint not null default 0,
   valid_until date not null,
   intro_note text,
   -- Staff-only. Never rendered into the guest email or the public page.
@@ -95,8 +96,8 @@ create table if not exists quote_items (
   ends_at timestamptz,
   rental_vehicle_slug text references rental_vehicles (slug),
   quantity int not null check (quantity > 0),
-  unit_amount_minor int not null check (unit_amount_minor >= 0),
-  subtotal_minor int not null check (subtotal_minor >= 0),
+  unit_amount_minor bigint not null check (unit_amount_minor >= 0),
+  subtotal_minor bigint not null check (subtotal_minor >= 0),
   -- A catalogue line must name an occurrence + option; a custom/rental line must not.
   constraint quote_item_shape check (
     (kind = 'catalogue' and session_occurrence_id is not null and activity_option_id is not null)
@@ -121,8 +122,8 @@ create table if not exists booking_custom_items (
   ends_at timestamptz,
   rental_vehicle_slug text references rental_vehicles (slug),
   quantity int not null check (quantity > 0),
-  unit_amount_minor int not null check (unit_amount_minor >= 0),
-  subtotal_minor int not null check (subtotal_minor >= 0),
+  unit_amount_minor bigint not null check (unit_amount_minor >= 0),
+  subtotal_minor bigint not null check (subtotal_minor >= 0),
   created_at timestamptz not null default now()
 );
 create index if not exists booking_custom_items_booking_idx on booking_custom_items (booking_id, position);
